Big update
New layout system in place, where the general content area now has the grid and containers deeper in the DOM, making it much easier to integrate full-width components (eg hero images) into non-full-width layouts.

Also, added a very lightweight bit of js to fix the FOUC (flash of unstyled content) when the page loads. This is in functions.php so that it can be run before the CSS gets loaded, and the class is added and removed via js to help prevent any problems with search engines etc. Finally, the class used only 'hides' the content visually, so screen readers can still access everything regardless.